requirements.txt
# Core
pandas
requests
yfinance
python-dotenv
streamlit
matplotlib>=3.0

# Broker integration
robin_stocks

# Optional exports and extensions
openpyxl

# Future scalability (for Phase 1.5+)
duckdb
scikit-learn

# Dev tools
black
flake8